backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.1
pydantic==2.6.1
openai==0.10.2  

# pip install -r requirements.txt